WebSockets/tests/submissions/Opera/interfaces/WebSocket/readyState/008.html

<!doctype html>
<title>WebSockets: getting readyState in closed</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../../constants.js></script>
<div id=log></div>
<script>
test(function(){
  var ws = new WebSocket(SCHEME_AND_DOMAIN+':80/ws/echo');
  ws.onopen = function(e) {
    debug(e);
    ws.onclose = function(e) {
      debug(e);
      assert_equals(ws.readyState, ws.CLOSED);
      end();
    }
    ws.close();
  }
  ws.onerror = ws.onmessage = ws.onclose = assertUnreached;
});
</script>
